requirements.txt
aiofiles==25.1.0
aiogram==3.23.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
Django==6.0
django-cors-headers==4.9.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
drf-yasg==1.21.11
frozenlist==1.8.0
idna==3.11
inflection==0.5.1
magic-filter==1.0.12
multidict==6.7.0
packaging==25.0
pillow==12.0.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
yarl==1.22.0
